Update build.yml
Revert to uploading extension artifacts because we can't add assets to a release before the release is created and the build workflow runs on commit to main. Anyhow, it will have the previous version until it gets bumped by jupyter_releaser in the prep-release workflow. I think we need to rebuild in the publish-release workflow so its artifacts will be accessible and have the correct version.